refactor: document components structure
