+++ frontend/src/components/documents/DocumentList/index.ts
@@ -1,0 +1,1 @@
+export { DocumentList } from './DocumentList';
