+++ frontend/src/components/documents/DocumentViewerPage/DocumentViewerPage.tsx
@@ -1,33 +1,44 @@
 import { RouteComponentProps } from 'react-router';
+
 import { PageLayout } from '../../common/PageLayout';
 import { Caption } from '../../common/Caption/Caption';
-import { useGetDocumentQuery } from '../../../generated/graphql';
-import { FileContentTextarea } from './FileContentTextArea';
 
-interface DocumentsDetailsPageProps
+import { useGetDocumentQuery } from '../../../generated/graphql';
+
+import { DocumentViewer } from '../DocumentViewer/DocumentViewer';
+
+interface DocumentViewerPageProps
   extends RouteComponentProps<{
     document_id: string;
   }> {}
 
-export const DocumentDetailsPage: React.FC<DocumentsDetailsPageProps> = ({
+export function DocumentViewerPage({
   match: {
     params: { document_id },
   },
-}: DocumentsDetailsPageProps) => {
+}: DocumentViewerPageProps) {
   const { data: documentData } = useGetDocumentQuery({
     variables: {
       document_id,
     },
   });
+
   if (!documentData?.getDocument.document) {
     return <PageLayout>Loading...</PageLayout>;
   }
-  const { file_name, file_url } = documentData.getDocument.document;
+
+  const { file_name } = documentData.getDocument.document;
 
   return (
     <PageLayout>
       <Caption>{file_name}</Caption>
-      <FileContentTextarea src={file_url} />
+      <DocumentViewer
+        documentId={document_id}
+        mode="view"
+        range={{}}
+        dots={[]}
+        onClickWord={() => {}}
+      />
     </PageLayout>
   );
-};
+}
