+++ frontend/src/components/documents/DocumentViewerPage/index.ts
@@ -1,0 +1,1 @@
+export { DocumentViewerPage } from './DocumentViewerPage';
