+++ frontend/src/components/documents/DocumentsPage/index.ts
@@ -1,0 +1,1 @@
+export { DocumentsPage } from './DocumentsPage';
